test: reorganize parser *_gaps.py files into topical modules
Distribute 9 meaningless *_gaps.py scratch files into proper topical test
modules that follow the existing test_arxml_parser_<topic>.py convention.

New shared infrastructure:
- conftest.py with reset_autosar/parser/warning_parser fixtures
  (R23-11 release pinned)
- _helpers.py with importable _snip() and _autosar_root() helpers

New topical modules:
- test_arxml_parser_e2e.py            (E2E Protection)
- test_arxml_parser_can_eth.py        (CAN FD, CouplingPort, Ethernet, VLAN)
- test_arxml_parser_nm_tp.py          (Network Management, Transport Protocol)
- test_arxml_parser_comspec.py        (Communication Specification)
- test_arxml_parser_multiplexed.py    (MultiplexedIPdu, SecureComm, SoAd)
- test_arxml_parser_diagnostics.py    (DoIp, DiagnosticConnection - new home)

Merges into existing files:
- swc_behavior_gaps, system_mapping_gaps -> orchestrators.py
- ecuc_values_gaps -> split: ecuc_handlers.py + network_handlers.py
- remaining_gaps (37 classes) -> distributed across 12 topical files

Existing destination files retain their local fixtures since their local
reset_autosar does not call setARRelease("R23-11"), unlike conftest. The
conftest serves the new files; existing files are untouched behaviorally.

Verified: 1243 parser tests collected (matches baseline), 4317 total pass,
flake8 clean, no test cases lost or duplicated. One class renamed to avoid
collision: TestReadRequiredComSpec -> TestReadRequiredComSpecExtras.

Closes #508
